update LuaGlue to the latest version
Former-commit-id: 40b2d74500c88bafc2ed3f0455737cc4c26230d1
Former-commit-id: 0965deb0597d01e8a787b633e6a6d6da22bf0d89
Former-commit-id: c73271748de88ecc324bbb54d0fbd15445f17a59

+++ plugins/lua/lib/LuaGlue/include/LuaGlue/LuaGlueMethod.h
@@ -55,7 +55,7 @@ class LuaGlueMethod : public LuaGlueMethodBase
 			ClassType *obj = *(ClassType **)lua_touserdata(state, 1);
 #endif
 			ReturnType ret = applyTuple(glueClass->luaGlue(), state, (_Class *)obj, fn, args);
-			if(Arg_Count_) lua_pop(state, Arg_Count_);
+			if(Arg_Count_) lua_pop(state, (int)Arg_Count_);
 			
 			returnValue(glueClass->luaGlue(), state, ret);
 			return 1;
